Step 2 completed
This update completes the code necessary for step two; there is a function to decompose the star-free emission line with default values (found through trial and error) for each line hard-coded into the function. Currently, function returns all gaussian components and their  parameters without identifying which curve corresponds to component, but it is unclear if this will be necessary. Values for alpha can be manually set as a tuple argument, and I suggest setting one as a very low value and the other as a higher value so as to create a high and low velocity curve fit. I have also included the option to perform a one phase decomposition, but I don'tt reccommend this. Further information on the specifics of the different decomposition parameters can be found in the documentation for GaussPy.