Paper Update
Final changes made to code for the purposes of the paper, no other major changes after this point.  Cleaned up some of the later code but decided that generalizing step 3 was unecessary given that there isn't analysis of other decompositions in the paper. Plots saved as .pgf files now, as Overleaf is able to render those with LaTeX and generate much cleaner plots compared to the .png files.  One bothersome side-effect of this is that matplotlib is unable to show plots until after code is run, so the plots are not returned alongside any verbose printed info but are returned after. Not a particularly pressing issue at this point since the code runs as intended.